+++ lib/src/common/exceptions/traced_exception.dart
@@ -1,0 +1,13 @@
+/// [exception] class for [TracedException].
+///
+/// Defines an exception with [StackTrace] information about where was thrown.
+final class TracedException implements Exception {
+  /// Exception data object.
+  final Object data;
+
+  /// Exception stack tracing information.
+  final StackTrace stackTrace;
+
+  /// Creates a new [TracedException] instance.
+  const TracedException(this.data, this.stackTrace);
+}