+++ lib/src/common/typedefs.dart
@@ -1,8 +1,8 @@
-/// Type definition for a [JSON] object.
-typedef JObject = Map<String, dynamic>;
+/// Type definition for a [DataMap] object.
+typedef DataMap = Map<String, Object?>;
 
-/// Type definition for a [JSON] entry.
-typedef JEntry = MapEntry<String, dynamic>;
+/// Type definition for a [DataMap] entry.
+typedef DataMapEntry = MapEntry<String, Object?>;
 
-/// Type definition for [HTTP] transaction headers.
-typedef CSMHeaders = Map<String, String>;
+/// Type definition for [HTTP] transaction [Headers].
+typedef Headers = Map<String, String>;