+++ lib/src/migrations/migrations_module.dart
@@ -1,2 +1,0 @@
-export 'interfaces/interfaces_module.dart';
-export 'structures/structures_module.dart';