+++ lib/src/service/entity/named_entity_b.dart
@@ -1,0 +1,12 @@
+import 'package:csm_client/csm_client.dart';
+
+/// [abstract] definition for [NamedEntities] simplification management.
+abstract class NamedEntityB extends EntityB implements NamedEntityI {
+  /// [Entity] unique name.
+  @override
+  String name = "";
+
+  /// [Entity] description.
+  @override
+  String? description;
+}