+++ lib/src/service/entity/named_entity_i.dart
@@ -1,0 +1,10 @@
+import 'package:csm_client/src/service/entity/entity_i.dart';
+
+/// [interface] definition for [entities] that are named.
+abstract interface class NamedEntityI implements EntityI {
+  /// [Entity] unique name.
+  late String name;
+
+  /// [Entity] description.
+  late String? description;
+}